requirements.txt
autopep8==1.7.0
colorama==0.4.5
cssbeautifier==1.14.6
email-validator==1.3.1
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-sqlalchemy==2.5.1
flask-wtf==1.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.12.0
pathspec==0.9.0
pyyaml==6.0
regex==2022.8.17
tomli==2.0.1
tqdm==4.64.0
